% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SailorMoon.R
\name{inthenameofthemoon}
\alias{inthenameofthemoon}
\title{A Sailor Moon palette generator}
\usage{
inthenameofthemoon(name, prism = 0)
}
\arguments{
\item{name}{Name of the Sailor Moon color palette you want to use. Use \code{MoonPaletteEnumeration} to see available options.}

\item{prism}{Use \code{prism = n} to interpolate a palette of n number of colors from the chosen palette. Useful if you want to use more color values than the palette provides. By default \code{prism} is set to 0 and will simply return the original palette.}
}
\value{
A vector of colors in hexadecimal notation.
}
\description{
Generates color palettes drawn from the (1992) Sailor Moon anime.
}
\examples{
inthenameofthemoon("MoonPrismPower")

inthenameofthemoon("MoonHealingEscalation", prism=16) #Returns the palette with 16 colors
}
\keyword{colors}
