% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SailorMoon.R
\docType{data}
\name{MoonPaletteEnumeration}
\alias{MoonPaletteEnumeration}
\title{List all Sailor Moon palettes}
\format{An object of class \code{character} of length 16.}
\usage{
MoonPaletteEnumeration
}
\description{
Gives a list of all Sailor Moon palettes available to use with \code{inthenameofthemoon()}.
}
\keyword{datasets}
