% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SailorMoon.R
\name{MoonPaletteIllumination}
\alias{MoonPaletteIllumination}
\title{Show a sample Sailor Moon palette}
\usage{
MoonPaletteIllumination(name)
}
\arguments{
\item{name}{Name of the Sailor Moon color palette you want to see a sample of. Run \code{MoonPaletteEnumeration} to see available options.}
}
\description{
Generates a swatch sample of an available Sailor Moon color palette.
}
\examples{
MoonPaletteIllumination("MoonPrincessHalation")
}
\keyword{colors}
