% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SailorMoon.R
\docType{data}
\name{MoonPalettes}
\alias{MoonPalettes}
\title{List of all palette names and their colors}
\format{An object of class \code{list} of length 16.}
\usage{
MoonPalettes
}
\description{
List of all palette names and their colors
}
\keyword{datasets}
