% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SailorMoon.R
\docType{data}
\name{sailorpalettes}
\alias{sailorpalettes}
\title{List all Sailor Moon palettes}
\format{An object of class \code{list} of length 4.}
\usage{
sailorpalettes
}
\description{
Gives a list of all Sailor Moon palettes available to use with \code{inthenameofthemoon()}.
}
\keyword{datasets}
