% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SailorMoon.R
\name{sailorswatch}
\alias{sailorswatch}
\title{Show a sample Sailor Moon palette}
\usage{
sailorswatch(name)
}
\arguments{
\item{name}{Name of the Sailor Moon color palette you want to see a sample of. Run \code{sailorpalettes} to see available options.}
}
\description{
Generates a swatch sample of an available Sailor Moon color palette.
}
\examples{
sailorswatch("MoonPrincessHalation")
}
\keyword{colors}
