% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SailorMoon.R
\name{inthenameofthemoon}
\alias{inthenameofthemoon}
\title{A Sailor Moon palette generator}
\usage{
inthenameofthemoon(name)
}
\arguments{
\item{name}{Name of the Sailor Moon color palette you want to use. Run \code{sailorpalettes} to see available options.}
}
\value{
A vector of colors in hex notation.
}
\description{
Generates color palettes drawn from the (1992) Sailor Moon anime.
}
\examples{
inthenameofthemoon("MoonPrismPower")
inthenameofthemoon("MoonHealingEscalation")
}
\keyword{colors}
